+++ factory/faker.py
@@ -45,15 +45,15 @@ class Faker(declarations.BaseDeclaration):
         self.provider_kwargs = kwargs
         self.locale = locale
 
-    def generate(self, extra_kwargs):
+    def generate(self, extra_kwargs=None):
         kwargs = {}
         kwargs.update(self.provider_kwargs)
-        kwargs.update(extra_kwargs)
+        kwargs.update(extra_kwargs or {})
         subfaker = self._get_faker(self.locale)
         return subfaker.format(self.provider, **kwargs)
 
     def evaluate(self, instance, step, extra):
-        return self.generate(extra or {})
+        return self.generate(extra)
 
     _FAKER_REGISTRY = {}
     _DEFAULT_LOCALE = faker.config.DEFAULT_LOCALE